experiments/field_matching/requirements.txt
pandas>=2.0.0
matplotlib>=3.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
rapidfuzz>=2.13.0
thefuzz>=0.19.0
jellyfish>=1.0.3
tqdm>=4.66.2 